requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
anthropic>=0.25.0
notion-client==2.2.1